NTC: ntc_phys: Add physical memory mapping library
This library supports DMA mapping of physical memory, so that it can be
accessed by the peer.  The peer may access the memory via a PCI-express
Non-Transparent Bridge (NTB), for example.  There may be other physical
transports besides NTB that require DMA mapping of physical memory.

A back end for NTB will be provided.  Another possiblity could be a back
end for SCIF devices, which are similar in nature to NTB and DMA, so
that NTC drivers can be used with Xeon Phi coprocessors.

Signed-off-by: Allen Hubbe <Allen.Hubbe@emc.com>