NTC: ntc_virt: Add virtual memory mapping library
This library supports kernel-virtual mapping of memory.  The peer may
access the memory via a TCP socket, for example.  The driver operating
the TCP socket may pass the kernel-virtual address of the memory as the
source or destination of kernel socket operations.  There may be
mechanisms other than kernel sockets that require kernel-virtual
mapping.

A TCP Non-Transparent Channel will be provided.  Another possibility for
NTC that would use virtual mapping would be file io.  Instead of a
socket, a file handle would be created for the NTC.  The file would then
be read and written in user space, redirected via netcat or ssh for
example, to connect to a remote peer, or output echoed to input to
create a loop back device.  The file input and output could also be
directed to predefined test inputs and checked against expected outputs,
for testing.

Signed-off-by: Allen Hubbe <Allen.Hubbe@emc.com>