NTRDMA: set remote resource key in init
If the remote resource key is not set before adding the remote resource,
then debugfs will not create a file name with the correct remote
resource key.  Since the key is decided by the local resource, we know
what key to use at remote resource initialization time, and we can set
the key then.

Signed-off-by: Allen Hubbe <Allen.Hubbe@emc.com>